Enable transfer learning
When we want to transfer learn from another model, we really only want
to initialise the weights from that model and nothing else.

Example usage:
  rave train --name xxx --db_path xxx --config /transfer/model/config.gin --config transfer --transfer_ckpt /transfer/model/checkpoint.ckpt

IMPORTANT(robin):

If you need to stop and restart a transfer learned model, you must only
pass the transfer learning flags the FIRST TIME YOU TRAIN THE MODEL.
This is because even if you pass --ckpt, the weights will be
re-initialised to the weights of the transfer model, effectively
resetting all of your progess.

The best way to handle this currently is to conditionally pass flags
using your shell, though I may change the code in the future to properly
ignore transfer_ckpt if an existing valid checkpoint is found.